Update GNJ to take a full distance matrix as input.
This is so the distance correction can be performed outside of GNJ
itself. The "similarity" matrix is still an input, so that the join
distances can be modulated based on the amount of data, but I'm not
actually sure this is worth the trouble and it may be best to add an
option turning that on/off to test if there is really an improvement.